extract out contest period component
